+++ 15-JPA/EjercicioAprendizajeExtra01/src/Libreria/persistencia/LibroJpaController.java
@@ -59,7 +59,7 @@ public class LibroJpaController implements Serializable {
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Long id = libro.getId();
+                Integer id = libro.getId();
                 if (findLibro(id) == null) {
                     throw new NonexistentEntityException("The libro with id " + id + " no longer exists.");
                 }
@@ -117,7 +117,7 @@ public class LibroJpaController implements Serializable {
         }
     }
 
-    public Libro findLibro(Long id) {
+    public Libro findLibro(Integer id) {
         EntityManager em = getEntityManager();
         try {
             return em.find(Libro.class, id);
@@ -137,6 +137,5 @@ public class LibroJpaController implements Serializable {
         } finally {
             em.close();
         }
-    }
-    
+    }    
 }
